grok_info_server/opengrok_api/requirement.txt
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
cached-property==1.5.1
channels==2.1.3
configparser==3.5.0
constantly==15.1.0
daphne==2.2.4
Django==2.0.9
django-cors-headers==2.4.0
django-secure==1.0.1
django-sslserver==0.20
django-url-filter==0.3.11
djangorestframework==3.9.0
djangosecure==0.0.11
enum-compat==0.0.2
future==0.16.0
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
pycryptodome==3.6.1
PyHamcrest==1.9.0
pytz==2018.7
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
zope.interface==4.6.0
